requirements.txt
# Core Dependencies
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
peft>=0.7.0
bitsandbytes>=0.41.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# NLP & Embeddings
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4  # Use faiss-gpu if CUDA available

# Prediction Models
catboost>=1.2.0
tabpfn>=0.1.8
shap>=0.43.0

# Validation & Metrics
rouge-score>=0.1.2
bert-score>=0.3.13

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
python-dotenv>=1.0.0

# Optional: For synthetic data generation
faker>=20.0.0
